+++ src/main/java/com/DisneyWorld/repo/IGeneroRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.DisneyWorld.model.Genero;
 
 @Repository
-public interface IGeneroRepo extends JpaRepository<Genero,Integer>{
+public interface IGeneroRepo extends JpaRepository<Genero,Long>{
 
 }
