+++ src/main/java/com/DisneyWorld/repo/IPersonajeRepo.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Repository;
 import com.DisneyWorld.model.Personaje;
 
 @Repository
-public interface IPersonajeRepo extends JpaRepository<Personaje,Integer>{
+public interface IPersonajeRepo extends JpaRepository<Personaje,Long>{
 
 	public List<Personaje> findAll();
-	public Personaje getById(Integer id);
+	public Personaje getById(Long id);
 	public Personaje save(Personaje personaje);
-	public void deleteById(Integer id);
+	public void deleteById(Long id);
 }
