+++ src/main/java/com/DisneyWorld/repo/IUsuarioRepo.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.DisneyWorld.model.Usuario;
 
 @Repository
-public interface IUsuarioRepo extends JpaRepository<Usuario,Integer> {
+public interface IUsuarioRepo extends JpaRepository<Usuario,Long> {
 
 		Usuario findByUsername(String username);
 }
